requirements.txt
python-dotenv
requests
numpy
pandas
plotly
glom
streamlit
folium
pillow
streamlit-folium
gunicorn
